rename: 기존 controller 테스트를 appceptance 테스트로 변경
